108 leave space (#734)
* [#108] WIP Begin static content of Leave Space form WIP


* [#108] WIP Delete space via client when Leave & Delete selected  WIP

Also reduce duplication in SpaceClient

* WIP [108] InviteEditors now works from SpaceDashboard

* [108] Add cypress test to validate going from dashboard tile to InviteEditors modal

* [#108] Implement back-end DELETE endpoint for Space objects

Requires owner permission; use with caution - cascades


* [108] Refresh space dashboard page after deleting a space

* [#108] Reassigning space ownership no longer cluttered with Share Link modal


* [#108] Style buttons to fit more text on Leave Space Modal

Also prevent canceling the modal from deleting your space LOL


* WIP [108] Don't show leave space for spaces that the user doens't own

* [108] Only show Leave Space if user owns the space.

Add mock to correct test run failures in SpaceDashboard

* [108] Add afterEach to clear jest mocks in an attempt to correct spurious failures due to test pollution

* [108] Add missing dependencies to get successful compilation of build

* [108] correct cypress text. extra call needed to be removed.

* [108] correct cypress text. Race condition (poorly) corrected.

* [108] Correct cypress test. Handle race condition in a more correct way

* [#108] Ådd new modal for transfer of space ownership


* [#108] Correct useEffect dependencies for TransferOwnershipForm


* [#108] Style changes only: basic layout for TransferOwnershipForm

* [#108] Begin adding radio buttons to Transfer Ownership form

* [#108] Assign New Owner form now looks as expected

* [#108] Spaces now refresh on the Space Dashboard after you leave one

* [#108] SpaceDashboardTile now only shows Leave Space for owners of spaces with editors

* [#108] clicking Leave Space on SpaceDashboardTile now goes directly to Assign New Owner form

This necessitated the test space Flipping Sweet be given Bob Barker as an editor, as lone users cannot assign new owners

* [#108] Add notification to the end of the Transfer Ownership workflow

Introduce a new type of NotificationModal for one-option messages

+++ api/src/main/kotlin/com/ford/internalprojects/peoplemover/space/SpaceService.kt
@@ -25,6 +25,7 @@ import com.ford.internalprojects.peoplemover.product.ProductService
 import com.ford.internalprojects.peoplemover.space.exceptions.SpaceIsReadOnlyException
 import com.ford.internalprojects.peoplemover.space.exceptions.SpaceNameInvalidException
 import com.ford.internalprojects.peoplemover.space.exceptions.SpaceNotExistsException
+import org.springframework.http.ResponseEntity
 import org.springframework.security.core.Authentication
 import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.stereotype.Service
@@ -87,8 +88,12 @@ class SpaceService(
         return spaceRepository.findByUuid(uuid) ?: throw SpaceNotExistsException()
     }
 
-    fun deleteSpace(uuid: String) {
-        spaceRepository.deleteByUuid(uuid)
+    fun deleteSpace(uuid: String) : Boolean {
+        if(spaceRepository.findByUuid(uuid) != null){
+            spaceRepository.deleteByUuid(uuid)
+            return true
+        }
+        return false
     }
 
     fun editSpace(uuid: String, editSpaceRequest: EditSpaceRequest): Space {
